Stability Kang Function

Performs a stability analysis based on the Kang (1988) criteria. Kang nonparametric stability (ranksum) uses both "trait single value" and stability variance (Shukla, 1972), and the genotype with the lowest ranksum is commonly the most favorable one.

Usage

stab_kang(.data, .y, .rep, .gen, .env)

# Default S3 method
stab_kang(.data, .y, .rep, .gen, .env)

Arguments

.data

data.frame

.y

Response Variable

.rep

Replication Factor

.gen

Genotypes Factor

.env

Environment Factor

References

Kang, M.S. (1988). A rank-sum method for selecting high-yielding, stable corn genotypes. Cereal Research Communications, 16, 1-2.

Shukla, G.K. (1972). Some aspects of partitioning genotype environmental components of variability. Heredity, 29, 237-245.

Author

Muhammad Yaseen (myaseen208@gmail.com)

Kent M. Edkridge (keskridge1@unl.edu)

Examples

data(ge_data)
YieldKang <-
     stab_kang(
            .data = ge_data
          , .y    = Yield
          , .rep  = Rep
          , .gen  = Gen
          , .env  = Env
      )
YieldKang
#> $Kang
#> # A tibble: 60 × 6
#>    Gen       Mean ShuklaVar rMean rShukaVar rStab
#>    <fct>    <dbl>     <dbl> <int>     <int> <int>
#>  1 013BT034 4248.    94993.    11         8    19
#>  2 122557   4062.   117014.    31        15    46
#>  3 122559   4259.   183294.    10        38    48
#>  4 12B.2511 4414.   255150.     4        52    56
#>  5 12FJ26   4057.   159191.    32        29    61
#>  6 14B.1030 3995.   135368.    37        22    59
#>  7 14C036   3943.   260565.    41        54    95
#>  8 14C040   3825.   148448.    53        27    80
#>  9 9496     4072.    99597.    30         9    39
#> 10 AUR0809  3792.   124782.    54        20    74
#> # ℹ 50 more rows
#>