Stability Analysis of Genotype by Environment Interaction (GEI)

The stability package provides functionalities to perform Stability Analysis of Genotype by Environment Interaction (GEI) to identify superior and stable genotypes under diverse environments. It performs Eberhart & Russel's ANOVA (1966), Finlay and Wilkinson (1963) Joint Linear Regression, Wricke (1962, 1964) Ecovalence, Shukla's stability variance parameter (1972) and Kang's (1991) simultaneous selection for high yielding and stable parameter.

Author

Muhammad Yaseen (myaseen208@gmail.com)

Kent M. Edkridge (keskridge1@unl.edu)